Move TestClientRegistrations to sagan
It will be used by additional tests and we want to avoid package tangles
