+++ sagan-site/src/test/java/sagan/TestClientRegistrations.java
@@ -1,4 +1,4 @@
-package sagan.security;
+package sagan;
 
 import org.springframework.security.oauth2.client.registration.ClientRegistration;
 import org.springframework.security.oauth2.core.AuthorizationGrantType;
