+++ sagan-site/src/test/java/sagan/security/GithubAuthenticationManagerTests.java
@@ -19,6 +19,7 @@ import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
 import org.springframework.security.oauth2.core.OAuth2Error;
 import org.springframework.security.oauth2.core.user.DefaultOAuth2User;
 import org.springframework.security.oauth2.core.user.OAuth2User;
+import sagan.TestClientRegistrations;
 
 import java.util.Collections;
