+++ sagan-site/src/test/java/sagan/security/GithubMemberOAuth2UserServiceTests.java
@@ -9,9 +9,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.client.registration.ClientRegistration;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import org.springframework.security.oauth2.core.AuthorizationGrantType;
-import org.springframework.security.oauth2.core.ClientAuthenticationMethod;
 import org.springframework.security.oauth2.core.OAuth2AccessToken;
+import sagan.TestClientRegistrations;
 
 import java.time.Duration;
 import java.time.Instant;
